+++ src/main/kotlin/team/jeonghokim/zakbu/global/exception/jwt/InvalidJwtException.kt
@@ -1,6 +1,6 @@
-package team.jeonghokim.zakbu.global.exception
+package team.jeonghokim.zakbu.global.exception.jwt
 
 import team.jeonghokim.zakbu.global.error.exception.ErrorCode
 import team.jeonghokim.zakbu.global.error.exception.ZakBuException
 
-object ExpiredJwtException: ZakBuException(ErrorCode.EXPIRED_JWT)
+object InvalidJwtException: ZakBuException(ErrorCode.INVALID_JWT)
