+++ src/main/kotlin/team/jeonghokim/zakbu/global/security/jwt/JwtTokenProvider.kt
@@ -9,8 +9,8 @@ import org.springframework.security.core.Authentication
 import org.springframework.security.core.userdetails.UserDetails
 import org.springframework.stereotype.Component
 import team.jeonghokim.zakbu.domain.auth.domain.RefreshToken
-import team.jeonghokim.zakbu.global.exception.ExpiredJwtException
-import team.jeonghokim.zakbu.global.exception.InvalidJwtException
+import team.jeonghokim.zakbu.global.exception.jwt.ExpiredJwtException
+import team.jeonghokim.zakbu.global.exception.jwt.InvalidJwtException
 import team.jeonghokim.zakbu.domain.auth.domain.repository.RefreshTokenRepository
 import team.jeonghokim.zakbu.global.security.auth.AuthDetailsService
 import java.util.*
